Implement clear player pref stuff.

+++ Tower-Style-Game/Assets/Scripts/Utils/PlayerPrefsCleaner.cs
@@ -2,8 +2,18 @@
 
 public class PlayerPrefsCleaner : MonoBehaviour {
 
+    [SerializeField]
+    private bool _forceDisableThisScript = false;
+
     private void Awake() {
-        PlayerPrefs.DeleteAll();
+        if (_forceDisableThisScript) {
+            this.enabled = false;
+            return;
+        }
+
+        if (Application.isEditor) {
+            PlayerPrefs.DeleteAll();
+        }
     }
 
 }
